requirements.txt
pymongo==3.10.1
python-decouple==3.3
Scrapy==2.1.0